lvl gap penalties, lvl sync and merits
mobcheck level difference based on highest lvl in party, then every member gets exp based on THEIR lvl for that same lvl gap between the monster and highest lvl in party. Level sync redundant so disabled. Merits changed to reflect lvl 75 exp and a !tnm command.

Known Issue: !tnm command is almost there but not quite working yet.

+++ sql/char_exp.sql
@@ -52,6 +52,6 @@ CREATE TABLE IF NOT EXISTS `char_exp` (
   `geo` int(10) unsigned NOT NULL DEFAULT '0',
   `run` int(10) unsigned NOT NULL DEFAULT '0',
   `merits` tinyint(2) unsigned NOT NULL DEFAULT '0',
-  `limits` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `limits` int(10) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`charid`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 AVG_ROW_LENGTH=85;
